Cast the settle window, which is the second one in the app and inherits the finding
Psalm's InvalidOperand at MotionService:641 — a `note`, so the job stays green and
the alert surfaces on the PR rather than failing the build. `microtime(true)` is a
float, an int/float division is `int|float`, and strict binary operands refuses the
mix.

RestoreService::staysListed() already carries exactly this cast with exactly this
reasoning; untrash()'s window is the second one in the app and copied the shape
without the cast. Cast rather than suppressed, so the arithmetic stays honest — a
settle of 6_000_000 really is six seconds, not zero.

+++ lib/Service/MotionService.php
@@ -638,7 +638,13 @@ final class MotionService {
 	 * is a final one.
 	 */
 	private function staysOutOfTheTrash(string $teamId, string $penpotId): bool {
-		$deadline = microtime(true) + (self::SETTLE_MICROSECONDS / 1_000_000.0);
+		// BOTH OPERANDS FLOAT, deliberately. `microtime(true)` is a float and an
+		// int/float division is `int|float`, which Psalm's strict binary operands
+		// mode refuses to mix. Casting rather than suppressing keeps the arithmetic
+		// honest — a settle of 6_000_000 really is six seconds, not zero.
+		// {@see RestoreService::staysListed()} carries the same cast for the same
+		// reason; this is the second window in the app and it inherits the finding.
+		$deadline = microtime(true) + ((float)self::SETTLE_MICROSECONDS / 1_000_000.0);
 		do {
 			usleep(self::SETTLE_POLL_MICROSECONDS);
